move sync to top level
